Fixed ish
i sort of got it working the timer sync is correct for all players now, along with the mouse catching display etc... only issue was all cats started in the water that messed it all up.  player scores are updating on the served for BOTH the mouse and the statue score so that means the only thing not working (again) is the display of that score.