Fixes
fixed the player freezing and unfreezing once all plyers join. Again still have the water rising issue and the score related issues. Next time I'm on, we focus on score, getting that fixed (look at new time sync) then implamenting the function that ends the game thic could also be in the Begin game script. that way the game loop is complete and all that needs done after is the fire cat tweaking and platform with the water + resetting the game/player health just the general stuff in my notes