Fixes
score is definitely working now, what a pain in the ass. but now the water ability istn working cats don't sink and water cat cant swim. however on the upside ive also added in the stoptimerRpc which should stop the clock and update the team score ifall players get 4 offerings to all 4 statues. still have to deal with health, resetting the game, network audio, and dealing with the leaderboard stuff. but ill begin all that once the main mechanics are fixed and working as they should.