Health system Done
health is set up along with the respawn points all that's left is to make sure that once at least one of the cats health hits 0 then the level conditions reset. either that or i can create a lose screen and then them back to the main menu where i already have the majority of the resetting logic in place.