Issue found and almost fixed
i was right all along, it is the network variables not resetting properly, now the game shutsdown simpler using the code i alreay previous had in the leave game function. and made all players reset their scenes after the game is shut down. and reset (almost all) network variables. i should be able to solve the issue by resetting the final network variables as well at the end or between the quitting of the game and scene reloading as i don't want to reference objects that have been dedstroyed.