Why am i doing this?
got the display to show when they all win, but the lose display only showed on the host therefore i need to fix or complete the network code for that. next time i'm on focus on completing the leader board stuff then see about fixing and finishing the reset before the deadline.